Add item in bulk api

+++ bistral-order-service/src/main/java/com/bistral/app/bistral_order_service/dtos/OrderItemRequest.java
@@ -23,6 +23,6 @@ public class OrderItemRequest {
     private UUID variantId;
     @NotNull
     @Min(1)
-    private int qty;
+    private Double qty;
 
 }
